add swagger , user and todo

+++ routes/api.php
@@ -18,25 +18,28 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+Route::get('/', [UserController::class, 'test']);
+
 //* Users */
 Route::get('/me', [UserController::class, 'me'])->middleware('auth:sanctum');
 Route::post('/login', [AuthController::class, 'login']);
 Route::post('/register', [AuthController::class, 'register']);
 Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
 Route::put('/users/{id}', [UserController::class, 'update'])->middleware('auth:sanctum');
-//Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
+Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
 
 //* Todo list */
-Route::get('/todo', [TodolistController::class, ''])->middleware('auth:sanctum');
-Route::post('/todo', [TodolistController::class, '']);
-Route::put('/todo/{id}', [TodolistController::class, '']);
-Route::delete('/todo/{id}', [TodolistController::class, '']);
+Route::get('/todolist', [TodolistController::class, 'todolist'])->middleware('auth:sanctum');
+Route::get('/todo/{id}', [TodolistController::class, 'todo'])->middleware('auth:sanctum');
+Route::post('/todo', [TodolistController::class, 'store'])->middleware('auth:sanctum');
+Route::put('/todo/{id}', [TodolistController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/todo/{id}', [TodolistController::class, 'destroy'])->middleware('auth:sanctum');
 
 //* Task */
-Route::get('/task', [TaskController::class, '']);
-Route::post('/task', [TaskController::class, '']);
-Route::put('/task', [TaskController::class, '']);
-Route::delete('/task', [TaskController::class, '']);
+Route::get('/task', [TaskController::class, ''])->middleware('auth:sanctum');
+Route::post('/task', [TaskController::class, ''])->middleware('auth:sanctum');
+Route::put('/task', [TaskController::class, ''])->middleware('auth:sanctum');
+Route::delete('/task', [TaskController::class, ''])->middleware('auth:sanctum');
 
 //* Category */
 // Route::get('/register', [CategoryController::class, '']);
